+++ crates/cgp-component/src/types/mod.rs
@@ -1,1 +1,3 @@
 pub mod delegate;
+
+pub use delegate::DelegateTo;
