+++ crates/cgp-core/Cargo.toml
@@ -14,6 +14,7 @@ description  = """
 [dependencies]
 cgp-async       = { version = "0.1.0" }
 cgp-component   = { version = "0.1.0" }
+cgp-type        = { version = "0.1.0" }
 cgp-error       = { version = "0.1.0" }
 cgp-field       = { version = "0.1.0" }
 cgp-inner       = { version = "0.1.0" }
