+++ crates/cgp-core/src/lib.rs
@@ -7,3 +7,4 @@ pub use cgp_component as component;
 pub use cgp_error as error;
 pub use cgp_field as field;
 pub use cgp_inner as inner;
+pub use cgp_type as types;
