+++ crates/cgp-core/src/prelude.rs
@@ -3,4 +3,4 @@ pub use cgp_component::{
     define_components, delegate_components, derive_component, DelegateComponent, HasComponents,
 };
 pub use cgp_error::{CanRaiseError, HasErrorType};
-pub use cgp_field::{symbol, Char, HasField};
+pub use cgp_field::{symbol, Char, HasField, HasFieldMut};
