+++ crates/cgp-type/src/impls/mod.rs
@@ -1,1 +1,3 @@
 pub mod use_type;
+
+pub use use_type::*;
