+++ crates/cgp-type/src/traits/mod.rs
@@ -1,1 +1,3 @@
-pub mod types;
+pub mod has_type;
+
+pub use has_type::*;
