+++ crates/cgp-type/src/types/mod.rs
@@ -1,1 +1,3 @@
 pub mod with_type;
+
+pub use with_type::WithTypeProvider;
